fix: abstraction of block classes
