+++ includes/Block/class-profile.php
@@ -12,36 +12,14 @@ defined( 'ABSPATH' ) || exit;
 /**
  * Register and handle the block.
  */
-class Profile {
+class Profile extends \Newspack\Govpack\Block {
 
 	/**
-	 * WordPress Hooks
-	 */
-	public static function hooks() {
-		add_action( 'init', [ __CLASS__, 'register' ] );
-	}
-
-	/**
-	 * Registers the block and associated script
+	 * Registers the block.
 	 *
 	 * @return void
 	 */
-	public static function register() {
-
-		$file = GOVPACK_PLUGIN_FILE . 'dist/editor.asset.php';
-
-		if ( file_exists( $file ) ) {
-			$asset_data = require_once $file; // phpcs:ignore WordPressVIPMinimum.Files.IncludingFile.UsingVariable
-		}
-
-		wp_register_script(
-			'govpack-editor',
-			plugin_dir_url( GOVPACK_PLUGIN_FILE ) . 'govpack/dist/editor.js',
-			$asset_data['dependencies'] ?? [],
-			$asset_data['version'] ?? '',
-			true
-		);
-
+	public static function register_block() {
 		register_block_type(
 			'govpack/profile',
 			[
